Split the router into one module per domain
`api/v1/routes.py` was 1701 lines: 25 endpoints across 9 domains, 14 private
helpers, and no way for two changes to touch it without meeting. It is now ten
modules, the largest 752 lines and the median 83.

Pure motion. Every function body is byte-identical to where it came from, and
the proof is not that the tests pass — it is that the generated OpenAPI schema
is byte-identical to `main`: 22 paths, every operation, parameter, response
model and tag unchanged. The 165 tests pass as well, and four import lines in
`test_notify.py` change only because the module they name no longer exists.

Three symbols stop being private, because a private name imported across a
package boundary is the worst of both: `_require` becomes `require_row` in a
new `_shared.py`, `_institution_out` becomes `institution_out` in `taxonomy`,
and `read_me` is now imported by name where `upsert_helper` answers with the
same body.

The prefix is declared once, in `api/v1/__init__.py`, so a module cannot
disagree with its neighbours about where it lives. `health` stays outside that
router: `/healthz` is what the deploy and the shared edge Caddy watch, and it
must not move when the API version does.

This is the first of five steps and it deliberately fixes nothing else. Eleven
endpoints still write to the database directly, there is still no transaction
rule, and `services/` still imports `api/schemas`. Doing any of that here
would have made this diff unreviewable, which is the whole reason the split
comes first.

`docs/architecture.md` is new and states the contract new code is held to,
with a closing section naming exactly where the code does not meet it yet — a
rule with silent exceptions is worse than no rule. README's palette paragraph
also moves the control to the header, where the previous change put it.

Docs updated first: docs/architecture.md, README.md
