+++ apps/api/src/konnekt/api/v1/__init__.py
@@ -1,0 +1,28 @@
+"""The versioned API, assembled from one module per domain.
+
+Every module here owns a slice of the URL space and nothing else: it parses the
+request, leans on `konnekt.services` for the rules, and renders a schema from
+`konnekt.api.schemas`. The prefix is declared once, here, so a module cannot
+disagree with its neighbours about where it lives.
+
+`health` is deliberately absent from this router — `/healthz` is not part of
+the versioned API and must not move when the version does.
+"""
+
+from fastapi import APIRouter
+
+from konnekt.api.v1 import (
+    browse,
+    cabinet,
+    me,
+    placements,
+    public,
+    requests,
+    search,
+    taxonomy,
+)
+
+router = APIRouter(prefix="/api/v1")
+
+for _module in (public, me, taxonomy, search, browse, cabinet, requests, placements):
+    router.include_router(_module.router)
