+++ apps/api/src/konnekt/api/v1/taxonomy.py
@@ -1,0 +1,211 @@
+"""Reference data — service types, subjects, institutions, languages.
+
+Read-only, translated in the database, and cached hard by the client. Nothing
+here belongs to a person.
+"""
+
+from fastapi import APIRouter, Query
+from sqlalchemy import func, select
+from sqlalchemy.orm import selectinload
+
+from konnekt.api.deps import LangDep, SessionDep, UserDep
+from konnekt.api.schemas import (
+    InstitutionOut,
+    LanguageOut,
+    ServiceTypeOut,
+    SubjectOut,
+)
+from konnekt.db.models import (
+    HelperProfile,
+    Institution,
+    Language,
+    Offer,
+    ServiceType,
+    Subject,
+)
+from konnekt.db.models.enums import (
+    PublishStatus,
+)
+from konnekt.services.catalog import _localised
+
+router = APIRouter()
+
+
+@router.get(
+    "/taxonomy/service-types", response_model=list[ServiceTypeOut], tags=["taxonomy"]
+)
+async def service_types(
+    session: SessionDep, lang: LangDep, user: UserDep
+) -> list[ServiceTypeOut]:
+    rows = (
+        await session.scalars(
+            select(ServiceType)
+            .where(ServiceType.is_active.is_(True))
+            .order_by(ServiceType.sort)
+            .options(selectinload(ServiceType.names))
+        )
+    ).all()
+    return [
+        ServiceTypeOut(
+            id=r.id,
+            code=r.code,
+            name=_localised(r.names, lang) or r.code,
+            hint=_localised(r.names, lang, "hint"),
+            requires_subject=r.requires_subject,
+            requires_institution=r.requires_institution,
+        )
+        for r in rows
+    ]
+
+
+@router.get("/taxonomy/subjects", response_model=list[SubjectOut], tags=["taxonomy"])
+async def subjects(
+    session: SessionDep,
+    lang: LangDep,
+    user: UserDep,
+    parent_id: int | None = None,
+    q: str | None = Query(default=None, max_length=200),
+    limit: int = Query(default=100, ge=1, le=300),
+) -> list[SubjectOut]:
+    """Browse the tree, or search it.
+
+    With `q` the answer comes from fuzzy matching, so "matan" and "матан" both
+    work. Without it, one level of the tree is returned.
+    """
+    if q:
+        from konnekt.services.lookup import find_subjects
+
+        matches = await find_subjects(session, q, lang, limit=limit, threshold=0.4)
+        ids = [m.id for m in matches]
+        if not ids:
+            return []
+        rows = {
+            s.id: s
+            for s in (
+                await session.scalars(
+                    select(Subject)
+                    .where(Subject.id.in_(ids))
+                    .options(selectinload(Subject.names))
+                )
+            ).all()
+        }
+        counts = await _offer_counts(session, ids)
+        return [
+            SubjectOut(
+                id=m.id,
+                slug=rows[m.id].slug,
+                name=m.label,
+                parent_id=rows[m.id].parent_id,
+                external_code=rows[m.id].external_code,
+                offers_count=counts.get(m.id, 0),
+            )
+            for m in matches
+            if m.id in rows
+        ]
+
+    stmt = (
+        select(Subject)
+        .where(Subject.is_active.is_(True), Subject.parent_id == parent_id)
+        .order_by(Subject.sort, Subject.id)
+        .options(selectinload(Subject.names))
+        .limit(limit)
+    )
+    rows = (await session.scalars(stmt)).all()
+    child_counts = {
+        parent_id: count
+        for parent_id, count in (
+            await session.execute(
+                select(Subject.parent_id, func.count(Subject.id))
+                .where(Subject.parent_id.in_([r.id for r in rows] or [0]))
+                .group_by(Subject.parent_id)
+            )
+        ).all()
+    }
+    counts = await _offer_counts(session, [r.id for r in rows])
+    return [
+        SubjectOut(
+            id=r.id,
+            slug=r.slug,
+            name=_localised(r.names, lang) or r.slug,
+            parent_id=r.parent_id,
+            has_children=child_counts.get(r.id, 0) > 0,
+            external_code=r.external_code,
+            offers_count=counts.get(r.id, 0),
+        )
+        for r in rows
+    ]
+
+
+async def _offer_counts(session, subject_ids: list[int]) -> dict[int, int]:
+    if not subject_ids:
+        return {}
+    rows = await session.execute(
+        select(Offer.subject_id, func.count(Offer.id))
+        .join(HelperProfile, HelperProfile.user_id == Offer.helper_id)
+        .where(
+            Offer.subject_id.in_(subject_ids),
+            Offer.is_active.is_(True),
+            HelperProfile.status == PublishStatus.PUBLISHED,
+        )
+        .group_by(Offer.subject_id)
+    )
+    return dict(rows.all())
+
+
+@router.get(
+    "/taxonomy/institutions", response_model=list[InstitutionOut], tags=["taxonomy"]
+)
+async def institutions(
+    session: SessionDep, lang: LangDep, user: UserDep
+) -> list[InstitutionOut]:
+    """The whole list, universities with their faculties nested.
+
+    A couple of hundred rows — cheaper to send once than to paginate, and the
+    client can then filter without another round trip.
+    """
+    rows = (
+        await session.scalars(
+            select(Institution)
+            .where(Institution.is_active.is_(True))
+            .order_by(Institution.sort, Institution.id)
+            .options(selectinload(Institution.names))
+        )
+    ).all()
+
+    by_parent: dict[int | None, list[Institution]] = {}
+    for row in rows:
+        by_parent.setdefault(row.parent_id, []).append(row)
+
+    return [
+        institution_out(row, lang, by_parent.get(row.id, []))
+        for row in by_parent.get(None, [])
+    ]
+
+
+def institution_out(row: Institution, lang, children=()) -> InstitutionOut:
+    return InstitutionOut(
+        id=row.id,
+        code=row.code,
+        name=_localised(row.names, lang) or row.code,
+        short_name=_localised(row.names, lang, "short_name"),
+        city=row.city,
+        parent_id=row.parent_id,
+        faculties=[institution_out(child, lang) for child in children],
+    )
+
+
+@router.get("/taxonomy/languages", response_model=list[LanguageOut], tags=["taxonomy"])
+async def languages(
+    session: SessionDep, lang: LangDep, user: UserDep
+) -> list[LanguageOut]:
+    rows = (
+        await session.scalars(
+            select(Language)
+            .where(Language.is_active.is_(True))
+            .order_by(Language.sort)
+            .options(selectinload(Language.names))
+        )
+    ).all()
+    return [
+        LanguageOut(code=r.code, name=_localised(r.names, lang) or r.code) for r in rows
+    ]
